updated to reduce usage with unneeded calls to extract expression values. updated to fix some bugs in the SQL queries preventing retrieving results.

+++ src/edu/ucdenver/ccp/PhenoGen/tools/analysis/GeneDataTools.java
@@ -833,6 +833,7 @@ public class GeneDataTools {
                 //if(!ucscComplete){
                 //       completedSuccessfully=false;
                 //}
+                log.debug("AsyncGeneDataTools with "+chrom+":"+minCoord+"-"+maxCoord);
                 prevThread=callAsyncGeneDataTools(chrom, minCoord, maxCoord,arrayTypeID,RNADatasetID);
                 /*boolean expError=callPanelExpr(chrom,minCoord,maxCoord,arrayTypeID,RNADatasetID,prevThread);
                 if(expError){
@@ -1889,129 +1890,135 @@ public class GeneDataTools {
         try{
             int publicUserID=new User().getUser_id("public",dbConn);
             String tmpoutputDir = fullPath + "tmpData/geneData/" + id + "/";
-            log.debug("createXML outputDir:"+tmpoutputDir);
-            File outDir=new File(tmpoutputDir);
-            if(outDir.exists()){
-                outDir.mkdirs();
-            }
-            Properties myProperties = new Properties();
-            File myPropertiesFile = new File(dbPropertiesFile);
-            myProperties.load(new FileInputStream(myPropertiesFile));
-
-            String dsn="dbi:"+myProperties.getProperty("PLATFORM") +":"+myProperties.getProperty("DATABASE");
-            String dbUser=myProperties.getProperty("USER");
-            String dbPassword=myProperties.getProperty("PASSWORD");
-
-            File ensPropertiesFile = new File(ensemblDBPropertiesFile);
-            Properties myENSProperties = new Properties();
-            myENSProperties.load(new FileInputStream(ensPropertiesFile));
-            String ensHost=myENSProperties.getProperty("HOST");
-            String ensPort=myENSProperties.getProperty("PORT");
-            String ensUser=myENSProperties.getProperty("USER");
-            String ensPassword=myENSProperties.getProperty("PASSWORD");
-            //construct perl Args
-            String[] perlArgs = new String[17];
-            perlArgs[0] = "perl";
-            perlArgs[1] = perlDir + "writeXML_RNA.pl";
-            perlArgs[2] = tmpoutputDir;
-            if (organism.equals("Rn")) {
-                perlArgs[3] = "Rat";
-            } else if (organism.equals("Mm")) {
-                perlArgs[3] = "Mouse";
-            }
-            perlArgs[4] = "Core";
-            perlArgs[5] = id;
-            perlArgs[6] = ucscDir+ucscGeneDir;
-            perlArgs[7] = Integer.toString(arrayTypeID);
-            perlArgs[8] = Integer.toString(rnaDS_ID);
-            perlArgs[9] = Integer.toString(publicUserID);
-            perlArgs[10] = dsn;
-            perlArgs[11] = dbUser;
-            perlArgs[12] = dbPassword;
-            perlArgs[13] = ensHost;
-            perlArgs[14] = ensPort;
-            perlArgs[15] = ensUser;
-            perlArgs[16] = ensPassword;
-            
-            
-            for (int i = 0; i < perlArgs.length; i++) {
-                log.debug(i + " PerlArgs::" + perlArgs[i]);
-                /*if(envVar[i].startsWith("PERL5LIB")&&organism.equals("Mm")){
-                    envVar[i]=envVar[i].replaceAll("ensembl_ucsc", "ensembl_ucsc_old");
-                }*/
-            }
-            
-            log.debug("setup params");
-            //set environment variables so you can access oracle pulled from perlEnvVar session variable which is a comma separated list
-            String[] envVar=perlEnvVar.split(",");
-
-            for (int i = 0; i < envVar.length; i++) {
-                log.debug(i + " EnvVar::" + envVar[i]);
-                /*if(envVar[i].startsWith("PERL5LIB")&&organism.equals("Mm")){
-                    envVar[i]=envVar[i].replaceAll("ensembl_ucsc", "ensembl_ucsc_old");
-                }*/
-            }
-
-            log.debug("setup envVar");
-            //construct ExecHandler which is used instead of Perl Handler because environment variables were needed.
-            myExec_session = new ExecHandler(perlDir, perlArgs, envVar, fullPath + "tmpData/geneData/"+id+"/");
-
-            try {
-
-                myExec_session.runExec();
-
-            } catch (ExecException e) {
-                completedSuccessfully=false;
-                log.error("In Exception of run findGeneRegion.pl Exec_session", e);
-                
-                String errorList=myExec_session.getErrors();
-                boolean missingDB=false;
-                String apiVer="";
-                
-                    if(errorList.contains("does not exist in DB.")){
-                        missingDB=true;
-                    }
-                    if(errorList.contains("Ensembl API version =")){
-                        int apiStart=errorList.indexOf("Ensembl API version =")+22;
-                        apiVer=errorList.substring(apiStart,apiStart+3);
-                    }
-                
-                if(!missingDB){
-                    setError("Running Perl Script to get Gene and Transcript details/images. Ensembl Assembly v"+apiVer);
-                }else{
-                    setError("The current Ensembl database does not have an entry for this gene ID."+
-                                " As Ensembl IDs are added/removed from new versions it is likely this ID has been removed."+
-                                " If you used a Gene Symbol and reached this the administrator will investigate. "+
-                                "If you entered this Ensembl ID please try to use a synonym or visit Ensembl to investigate the status of this ID. "+
-                                "Ensembl Assembly v"+apiVer);
-                                        
+            File test=new File(tmpoutputDir+"Region.xml");
+            long testLM=test.lastModified();
+            testLM=(new Date().getTime())-testLM;
+            long fifteenMin=15*60*1000;
+            if(!test.exists() || (test.length()==0 && testLM>fifteenMin)){
+                log.debug("createXML outputDir:"+tmpoutputDir);
+                File outDir=new File(tmpoutputDir);
+                if(outDir.exists()){
+                    outDir.mkdirs();
                 }
-                
-                Email myAdminEmail = new Email();
-                myAdminEmail.setSubject("Exception thrown in Exec_session");
-                myAdminEmail.setContent("There was an error while running "
-                        + perlArgs[1] + " (" + perlArgs[2] +" , "+perlArgs[3]+" , "+perlArgs[4]+" , "+perlArgs[5]+" , "+perlArgs[6]+","+perlArgs[7]+
-                        ")\n\n"+myExec_session.getErrors());
-                try {
-                    myAdminEmail.sendEmailToAdministrator((String) session.getAttribute("adminEmail"));
-                } catch (Exception mailException) {
-                    log.error("error sending message", mailException);
-                    throw new RuntimeException();
-                }
-            }
+                Properties myProperties = new Properties();
+                File myPropertiesFile = new File(dbPropertiesFile);
+                myProperties.load(new FileInputStream(myPropertiesFile));
 
-            if(myExec_session.getExitValue()!=0){
-                completedSuccessfully=false;
-                Email myAdminEmail = new Email();
-                myAdminEmail.setSubject("Exception thrown in Exec_session");
-                myAdminEmail.setContent("There was an error while running "
-                        + perlArgs[1] + " (" + perlArgs[2] +" , "+perlArgs[3]+" , "+perlArgs[4]+" , "+perlArgs[5]+" , "+perlArgs[6]+
-                        ")\n\n"+myExec_session.getErrors());
+                String dsn="dbi:"+myProperties.getProperty("PLATFORM") +":"+myProperties.getProperty("DATABASE");
+                String dbUser=myProperties.getProperty("USER");
+                String dbPassword=myProperties.getProperty("PASSWORD");
+
+                File ensPropertiesFile = new File(ensemblDBPropertiesFile);
+                Properties myENSProperties = new Properties();
+                myENSProperties.load(new FileInputStream(ensPropertiesFile));
+                String ensHost=myENSProperties.getProperty("HOST");
+                String ensPort=myENSProperties.getProperty("PORT");
+                String ensUser=myENSProperties.getProperty("USER");
+                String ensPassword=myENSProperties.getProperty("PASSWORD");
+                //construct perl Args
+                String[] perlArgs = new String[17];
+                perlArgs[0] = "perl";
+                perlArgs[1] = perlDir + "writeXML_RNA.pl";
+                perlArgs[2] = tmpoutputDir;
+                if (organism.equals("Rn")) {
+                    perlArgs[3] = "Rat";
+                } else if (organism.equals("Mm")) {
+                    perlArgs[3] = "Mouse";
+                }
+                perlArgs[4] = "Core";
+                perlArgs[5] = id;
+                perlArgs[6] = ucscDir+ucscGeneDir;
+                perlArgs[7] = Integer.toString(arrayTypeID);
+                perlArgs[8] = Integer.toString(rnaDS_ID);
+                perlArgs[9] = Integer.toString(publicUserID);
+                perlArgs[10] = dsn;
+                perlArgs[11] = dbUser;
+                perlArgs[12] = dbPassword;
+                perlArgs[13] = ensHost;
+                perlArgs[14] = ensPort;
+                perlArgs[15] = ensUser;
+                perlArgs[16] = ensPassword;
+
+
+                for (int i = 0; i < perlArgs.length; i++) {
+                    log.debug(i + " PerlArgs::" + perlArgs[i]);
+                    /*if(envVar[i].startsWith("PERL5LIB")&&organism.equals("Mm")){
+                        envVar[i]=envVar[i].replaceAll("ensembl_ucsc", "ensembl_ucsc_old");
+                    }*/
+                }
+
+                log.debug("setup params");
+                //set environment variables so you can access oracle pulled from perlEnvVar session variable which is a comma separated list
+                String[] envVar=perlEnvVar.split(",");
+
+                for (int i = 0; i < envVar.length; i++) {
+                    log.debug(i + " EnvVar::" + envVar[i]);
+                    /*if(envVar[i].startsWith("PERL5LIB")&&organism.equals("Mm")){
+                        envVar[i]=envVar[i].replaceAll("ensembl_ucsc", "ensembl_ucsc_old");
+                    }*/
+                }
+
+                log.debug("setup envVar");
+                //construct ExecHandler which is used instead of Perl Handler because environment variables were needed.
+                myExec_session = new ExecHandler(perlDir, perlArgs, envVar, fullPath + "tmpData/geneData/"+id+"/");
+
                 try {
-                    myAdminEmail.sendEmailToAdministrator((String) session.getAttribute("adminEmail"));
-                } catch (Exception mailException) {
-                    log.error("error sending message", mailException);
-                    throw new RuntimeException();
+
+                    myExec_session.runExec();
+
+                } catch (ExecException e) {
+                    completedSuccessfully=false;
+                    log.error("In Exception of run findGeneRegion.pl Exec_session", e);
+
+                    String errorList=myExec_session.getErrors();
+                    boolean missingDB=false;
+                    String apiVer="";
+
+                        if(errorList.contains("does not exist in DB.")){
+                            missingDB=true;
+                        }
+                        if(errorList.contains("Ensembl API version =")){
+                            int apiStart=errorList.indexOf("Ensembl API version =")+22;
+                            apiVer=errorList.substring(apiStart,apiStart+3);
+                        }
+
+                    if(!missingDB){
+                        setError("Running Perl Script to get Gene and Transcript details/images. Ensembl Assembly v"+apiVer);
+                    }else{
+                        setError("The current Ensembl database does not have an entry for this gene ID."+
+                                    " As Ensembl IDs are added/removed from new versions it is likely this ID has been removed."+
+                                    " If you used a Gene Symbol and reached this the administrator will investigate. "+
+                                    "If you entered this Ensembl ID please try to use a synonym or visit Ensembl to investigate the status of this ID. "+
+                                    "Ensembl Assembly v"+apiVer);
+
+                    }
+
+                    Email myAdminEmail = new Email();
+                    myAdminEmail.setSubject("Exception thrown in Exec_session");
+                    myAdminEmail.setContent("There was an error while running "
+                            + perlArgs[1] + " (" + perlArgs[2] +" , "+perlArgs[3]+" , "+perlArgs[4]+" , "+perlArgs[5]+" , "+perlArgs[6]+","+perlArgs[7]+
+                            ")\n\n"+myExec_session.getErrors());
+                    try {
+                        myAdminEmail.sendEmailToAdministrator((String) session.getAttribute("adminEmail"));
+                    } catch (Exception mailException) {
+                        log.error("error sending message", mailException);
+                        throw new RuntimeException();
+                    }
+                }
+
+                if(myExec_session.getExitValue()!=0){
+                    completedSuccessfully=false;
+                    Email myAdminEmail = new Email();
+                    myAdminEmail.setSubject("Exception thrown in Exec_session");
+                    myAdminEmail.setContent("There was an error while running "
+                            + perlArgs[1] + " (" + perlArgs[2] +" , "+perlArgs[3]+" , "+perlArgs[4]+" , "+perlArgs[5]+" , "+perlArgs[6]+
+                            ")\n\n"+myExec_session.getErrors());
+                    try {
+                        myAdminEmail.sendEmailToAdministrator((String) session.getAttribute("adminEmail"));
+                    } catch (Exception mailException) {
+                        log.error("error sending message", mailException);
+                        throw new RuntimeException();
+                    }
                 }
             }
         }catch(Exception e){
@@ -2056,35 +2063,43 @@ public class GeneDataTools {
                 //log.debug("Getting ready to start");
                 File indivf=new File(tmpOutput+"Panel_Expr_indiv_tmp.txt");
                 File groupf=new File(tmpOutput+"Panel_Expr_group_tmp.txt");
-                BufferedWriter outGroup=new BufferedWriter(new FileWriter(groupf));
-                BufferedWriter outIndiv=new BufferedWriter(new FileWriter(indivf));
-                ArrayList<AsyncGeneDataExpr> localList=new ArrayList<AsyncGeneDataExpr>();
-                SyncAndClose sac=new SyncAndClose(start,localList,dbConn,outGroup,outIndiv,usageID,outputDir);
-                while(rs.next()){
-                    AsyncGeneDataExpr agde=new AsyncGeneDataExpr(session,tmpOutput+"tmp_psList.txt",tmpOutput,null,threadList,maxThreadRunning,outGroup,outIndiv,sac,ver);
-                    String dataset_id=Integer.toString(rs.getInt("DATASET_ID"));
-                    int iDSID=rs.getInt("DATASET_ID");
-                    String tissue=rs.getString("TISSUE");
-                    String tissueNoSpaces=tissue.replaceAll(" ", "_");
-                    edu.ucdenver.ccp.PhenoGen.data.Dataset sDataSet=new edu.ucdenver.ccp.PhenoGen.data.Dataset();
-                    edu.ucdenver.ccp.PhenoGen.data.Dataset curDS=sDataSet.getDataset(iDSID,dbConn,"");
-                    String affyFile="allPS";
-                    String verStr="allPS";
-                    if(arrayTypeID==21){
-                        affyFile="NormVer";
-                        verStr=ver;
+                long curTime=new Date().getTime();
+                long indLM=indivf.lastModified();
+                long groupLM=groupf.lastModified();
+                indLM=curTime-indLM;
+                groupLM=curTime-groupLM;
+                long twoHours=1000*60*60*2;
+                if(!indivf.exists() || !groupf.exists()||((groupf.length()==0 && groupLM>twoHours) || (indivf.length()==0 && indLM>twoHours))){
+                    BufferedWriter outGroup=new BufferedWriter(new FileWriter(groupf));
+                    BufferedWriter outIndiv=new BufferedWriter(new FileWriter(indivf));
+                    ArrayList<AsyncGeneDataExpr> localList=new ArrayList<AsyncGeneDataExpr>();
+                    SyncAndClose sac=new SyncAndClose(start,localList,dbConn,outGroup,outIndiv,usageID,outputDir);
+                    while(rs.next()){
+                        AsyncGeneDataExpr agde=new AsyncGeneDataExpr(session,tmpOutput+"tmp_psList.txt",tmpOutput,null,threadList,maxThreadRunning,outGroup,outIndiv,sac,ver);
+                        String dataset_id=Integer.toString(rs.getInt("DATASET_ID"));
+                        int iDSID=rs.getInt("DATASET_ID");
+                        String tissue=rs.getString("TISSUE");
+                        String tissueNoSpaces=tissue.replaceAll(" ", "_");
+                        edu.ucdenver.ccp.PhenoGen.data.Dataset sDataSet=new edu.ucdenver.ccp.PhenoGen.data.Dataset();
+                        edu.ucdenver.ccp.PhenoGen.data.Dataset curDS=sDataSet.getDataset(iDSID,dbConn,"");
+                        String affyFile="allPS";
+                        String verStr="allPS";
+                        if(arrayTypeID==21){
+                            affyFile="NormVer";
+                            verStr=ver;
+                        }
+                        String DSPath=userFilesRoot+"public/Datasets/"+curDS.getNameNoSpaces()+"_Master/Affy."+affyFile+".h5";
+                        String sampleFile=userFilesRoot+"public/Datasets/"+curDS.getNameNoSpaces()+"_Master/"+verStr+"_samples.txt";
+                        String groupFile=userFilesRoot+"public/Datasets/"+curDS.getNameNoSpaces()+"_Master/"+verStr+"_groups.txt";
+                        String outGroupFile="group_"+tissueNoSpaces+"_exprVal.txt";
+                        String outIndivFile="indiv_"+tissueNoSpaces+"_exprVal.txt";
+                        agde.add(DSPath,sampleFile,groupFile,outGroupFile,outIndivFile,tissue,curDS.getPlatform());
+                        threadList.add(agde);
+                        localList.add(agde);
+                        agde.start();     
                     }
-                    String DSPath=userFilesRoot+"public/Datasets/"+curDS.getNameNoSpaces()+"_Master/Affy."+affyFile+".h5";
-                    String sampleFile=userFilesRoot+"public/Datasets/"+curDS.getNameNoSpaces()+"_Master/"+verStr+"_samples.txt";
-                    String groupFile=userFilesRoot+"public/Datasets/"+curDS.getNameNoSpaces()+"_Master/"+verStr+"_groups.txt";
-                    String outGroupFile="group_"+tissueNoSpaces+"_exprVal.txt";
-                    String outIndivFile="indiv_"+tissueNoSpaces+"_exprVal.txt";
-                    agde.add(DSPath,sampleFile,groupFile,outGroupFile,outIndivFile,tissue,curDS.getPlatform());
-                    threadList.add(agde);
-                    localList.add(agde);
-                    agde.start();     
+                    ps.close();
                 }
-                ps.close();
                 
                 //log.debug("Started AsyncGeneDataExpr");
             }catch(IOException ioe){
